update post service
