+++ backend/src/modules/community/community.controller.ts
@@ -3,7 +3,6 @@ import { CommunityService } from './community.service';
 import { AddUserToCommunityDto, CreateCommunityDto } from './dto/community.dto';
 import { ResponseHelper, ResponseInterface } from 'src/helper/response.helper';
 import { JwtAuthGuard } from 'src/guards/jwt-auth.guard';
-import { use } from 'passport';
 
 @Controller('community')
 export class CommunityController {
@@ -94,55 +93,6 @@ export class CommunityController {
   async fetchAllCommunitiesWithUsersStatus(@Param('user_id') userId: string) {
     try {
       const res: ResponseInterface = await this.communityService.fetchAllCommunitiesWithUsersStatus(userId);
-      return ResponseHelper.successResponse(res.message, res.status_code, res.data)
-    } catch(error) {
-      console.log(error)
-      throw new HttpException(error.message, error.status || HttpStatus.INTERNAL_SERVER_ERROR)   
-    }
-  }
-
-  // posts
-  @UseGuards(JwtAuthGuard)
-  @Get('/posts/')
-  async fetchAllPosts() {
-    try {
-      const res: ResponseInterface = await this.communityService.fetchAllPosts();
-      return ResponseHelper.successResponse(res.message, res.status_code, res.data)
-    } catch(error) {
-      console.log(error)
-      throw new HttpException(error.message, error.status || HttpStatus.INTERNAL_SERVER_ERROR)   
-    }
-  }
-
-  @UseGuards(JwtAuthGuard)
-  @Get('/posts/:id')
-  async fetchPostsById(@Param('id') userId: string) {
-    try {
-      const res: ResponseInterface = await this.communityService.fetchPostById(userId);
-      return ResponseHelper.successResponse(res.message, res.status_code, res.data)
-    } catch(error) {
-      console.log(error)
-      throw new HttpException(error.message, error.status || HttpStatus.INTERNAL_SERVER_ERROR)   
-    }
-  }
-
-  @UseGuards(JwtAuthGuard)
-  @Get('/:community_id/posts')
-  async fetchPostsByCommunity(@Param('community_id') communityId: string) {
-    try {
-      const res: ResponseInterface = await this.communityService.fetchPostsByCommunity(communityId);
-      return ResponseHelper.successResponse(res.message, res.status_code, res.data)
-    } catch(error) {
-      console.log(error)
-      throw new HttpException(error.message, error.status || HttpStatus.INTERNAL_SERVER_ERROR)   
-    }
-  }
-
-  @UseGuards(JwtAuthGuard)
-  @Get('/user/:user_id/posts/')
-  async fetchPostsByUser(@Param('user_id') userId: string) {
-    try {
-      const res: ResponseInterface = await this.communityService.fetchPostsByUser(userId);
       return ResponseHelper.successResponse(res.message, res.status_code, res.data)
     } catch(error) {
       console.log(error)
