+++ backend/src/modules/community/community.service.ts
@@ -6,7 +6,7 @@ import { Repository } from 'typeorm';
 import { CreateCommunityInterface } from './interfaces/community.interfaces';
 import { ResponseInterface } from 'src/helper/response.helper';
 import { UserService } from '../user/user.service';
-import { PostService } from '../post/post.service';
+
 @Injectable()
 export class CommunityService {
   constructor(
@@ -17,7 +17,6 @@ export class CommunityService {
     private userCommunityRepo: Repository<UserCommunity>,
 
     private userService: UserService,
-    private postService: PostService,
   ){}
 
   async createCommunity(createCommunityInterface: CreateCommunityInterface): Promise<ResponseInterface> {
@@ -178,63 +177,4 @@ export class CommunityService {
     }
   }
 
-  async fetchPostById(postId: string) {
-    const existingPost = await this.postService.fetchPostById(postId);
-
-    return {
-      status: "success",
-      status_code: 200,
-      message: "Post fetched successfully",
-      data: existingPost
-    }
-  }
-
-  async fetchAllPosts() {
-    const posts = await this.postService.fetchAllPosts();
-
-    return {
-      status: "success",
-      status_code: 200,
-      message: "Posts fetched successfully",
-      data: posts
-    }
-  }
-
-  async fetchPostsByCommunity(communityId: string) {
-    const existingCommunity = await this.communityRepo.findOne({
-      where: {
-        id: communityId
-      }
-    });
-
-    if (!existingCommunity) {
-      throw new HttpException("Community Id is not valid", HttpStatus.BAD_REQUEST);
-    }
-
-    const posts = await this.postService.fetchPostsByCommunity(communityId)
-
-    return {
-      status: "success",
-      status_code: 200,
-      message: "Posts retrieved successfully",
-      data: posts
-    }
-  }
-
-  async fetchPostsByUser(userId: string) {
-    const existingUser = await this.userService.findUserById(userId);
-
-    if (!existingUser) {
-      throw new HttpException("User Id is not valid", HttpStatus.BAD_REQUEST);
-    }
-
-    const posts = await this.postService.fetchPostsByUser(userId)
-
-    return {
-      status: "success",
-      status_code: 200,
-      message: "Posts retrieved successfully",
-      data: posts
-    }
-  }
 }
