+++ backend/src/modules/post/post.module.ts
@@ -4,13 +4,14 @@ import { Post } from 'src/entities/posts.entity';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { PostController } from './post.controller';
 import { Community } from 'src/entities/community.entity';
+import { UserModule } from '../user/user.module';
 
 @Module({
   imports: [
-    TypeOrmModule.forFeature([Post, Community])
+    TypeOrmModule.forFeature([Post, Community]),
+    UserModule,
   ],
   providers: [PostService],
-  exports: [PostService],
   controllers: [PostController]
 })
 export class PostModule {}
